+++ lib/kitchen/directions/move_solutions_to_answer_key/move_solutions_from_numbered_note.rb
@@ -1,0 +1,33 @@
+# frozen_string_literal: true
+
+module Kitchen::Directions::MoveSolutionsFromNumberedNote
+  def self.v1(chapter:, append_to:, note_class:)
+    V1.new.bake(chapter: chapter, append_to: append_to, note_class: note_class)
+  end
+
+  class V1
+    def bake(chapter:, append_to:, note_class:)
+      solutions_clipboard = Kitchen::Clipboard.new
+      chapter.notes("$.#{note_class}").exercises.each do |exercise|
+        solution = exercise.solution
+        next unless solution
+
+        solution.cut(to: solutions_clipboard)
+      end
+
+      return if solutions_clipboard.items.empty?
+
+      title = <<~HTML
+        <h3 data-type="title">
+          <span class="os-title-label">#{I18n.t(:"notes.#{note_class}")}</span>
+        </h3>
+      HTML
+
+      append_to.append(child:
+        Kitchen::Directions::SolutionAreaSnippet.v1(
+          title: title, solutions_clipboard: solutions_clipboard
+        )
+      )
+    end
+  end
+end