+++ lib/kitchen/directions/move_solutions_to_answer_key/solution_area_snippet.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module Kitchen::Directions::SolutionAreaSnippet
+  def self.v1(title:, solutions_clipboard:)
+    <<~HTML
+      <div class="os-solution-area">
+        #{title}
+        #{solutions_clipboard.paste}
+      </div>
+    HTML
+  end
+end