+++ lib/kitchen/directions/move_solutions_to_answer_key/strategies/contemporary_math.rb
@@ -4,36 +4,19 @@ module Kitchen::Directions::MoveSolutionsToAnswerKey
   module Strategies
     class ContemporaryMath
       def bake(chapter:, append_to:)
-        solutions_clipboard = Kitchen::Clipboard.new
+        # Hacky numbering fix
         chapter.notes('$.your-turn').exercises.each do |exercise|
           solution = exercise.solution
           next unless solution
 
-          # Hacky numbering fix
           number = exercise.ancestor(:note).count_in(:chapter)
           solution.first('a.os-number').inner_html = number.to_s
           solution.first('span.os-divider').inner_html = '. '
-          solution.cut(to: solutions_clipboard)
         end
-        title = <<~HTML
-          <h3 data-type="title">
-            <span class="os-title-label">#{I18n.t(:'notes.your-turn')}</span>
-          </h3>
-        HTML
-        append_solution_area(title: title, solutions: solutions_clipboard,
-                             append_to: append_to)
-      end
 
-      def append_solution_area(title:, solutions:, append_to:)
-        append_to = append_to.add_child(
-          <<~HTML
-            <div class="os-solution-area">
-              #{title}
-            </div>
-          HTML
-        ).first
-
-        append_to.add_child(solutions.paste)
+        Kitchen::Directions::MoveSolutionsFromNumberedNote.v1(
+          chapter: chapter, append_to: append_to, note_class: 'your-turn'
+        )
       end
     end
   end